#50 [FIX] service 순환참조 에러로 CustomProvider 추가 후 메소드 이전
